+++ packages/frontend-2/components/projects/Dashboard.vue
@@ -1,28 +1,10 @@
 <template>
   <div>
     <Portal to="primary-actions"></Portal>
-    <div
-      class="w-[calc(100vw-8px)] ml-[calc(50%-50vw+4px)] mr-[calc(50%-50vw+4px)] -mt-6 mb-10 bg-foundation divide-y divide-outline-3 border-b border-outline-3"
-    >
-      <div v-if="showChecklist">
-        <OnboardingChecklistV1 show-intro />
-      </div>
-      <ProjectsInviteBanners
-        v-if="projectsPanelResult?.activeUser?.projectInvites?.length"
-        :invites="projectsPanelResult?.activeUser"
-      />
-      <WorkspaceInviteBanners
-        v-if="projectsPanelResult?.activeUser?.workspaceInvites?.length"
-        :invites="projectsPanelResult.activeUser"
-      >
-        <ProjectsNewSpeckleBanner
-          v-if="showNewSpeckleBanner"
-          @dismissed="onDismissNewSpeckleBanner"
-        />
-      </WorkspaceInviteBanners>
-    </div>
-
-    <PromoBannersWrapper v-if="promoBanners.length" :banners="promoBanners" />
+    <ProjectsDashboardHeader
+      :user="projectsPanelResult?.activeUser || undefined"
+      class="mb-10"
+    />
 
     <div v-if="!showEmptyState" class="flex flex-col gap-4">
       <div class="flex items-center gap-2 mb-2">
@@ -92,8 +74,6 @@ import { ToastNotificationType, useGlobalToast } from '~~/lib/common/composables
 import { projectRoute } from '~~/lib/common/helpers/route'
 import { useActiveUser } from '~~/lib/auth/composables/activeUser'
 import type { Nullable, Optional, StreamRoles } from '@speckle/shared'
-import { useSynchronizedCookie } from '~~/lib/common/composables/reactiveCookie'
-import type { PromoBanner } from '~/lib/promo-banners/types'
 import { useDebouncedTextInput, type InfiniteLoaderState } from '@speckle/ui-components'
 import { MagnifyingGlassIcon, Squares2X2Icon } from '@heroicons/vue/24/outline'
 
@@ -108,17 +88,6 @@ const { activeUser, isGuest } = useActiveUser()
 const { triggerNotification } = useGlobalToast()
 const areQueriesLoading = useQueryLoading()
 const apollo = useApolloClient().client
-
-const promoBanners = ref<PromoBanner[]>([
-  {
-    id: 'speckleverse',
-    primaryText: 'Join our online hackathon!',
-    secondaryText: 'June 7 - 9, 2024',
-    url: 'https://beyond-the-speckleverse.devpost.com/',
-    priority: 1,
-    expiryDate: '2024-06-10'
-  }
-])
 
 const {
   on,
@@ -272,59 +241,5 @@ watch(areQueriesLoading, (newVal) => (showLoadingBar.value = newVal))
 const clearSearch = () => {
   search.value = ''
   selectedRoles.value = []
-}
-
-const hasCompletedChecklistV1 = useSynchronizedCookie<boolean>(
-  `hasCompletedChecklistV1`,
-  {
-    default: () => false
-  }
-)
-
-const hasDismissedChecklistTime = useSynchronizedCookie<string | undefined>(
-  `hasDismissedChecklistTime`,
-  { default: () => undefined }
-)
-
-const hasDismissedChecklistForever = useSynchronizedCookie<boolean | undefined>(
-  `hasDismissedChecklistForever`,
-  {
-    default: () => false
-  }
-)
-
-const hasDismissedChecklistTimeAgo = computed(() => {
-  return (
-    new Date().getTime() -
-    new Date(hasDismissedChecklistTime.value || Date.now()).getTime()
-  )
-})
-
-const hasDismissedNewSpeckleBanner = useSynchronizedCookie<boolean | undefined>(
-  `hasDismissedNewSpeckleBanner`,
-  { default: () => false }
-)
-
-const showChecklist = computed(() => {
-  if (hasDismissedChecklistForever.value) return false
-  if (hasCompletedChecklistV1.value) return false
-  if (hasDismissedChecklistTime.value === undefined) return true
-  if (
-    hasDismissedChecklistTime.value !== undefined &&
-    hasDismissedChecklistTimeAgo.value > 86400000
-  )
-    return true
-  return false
-})
-
-const showNewSpeckleBanner = computed(() => {
-  if (hasDismissedNewSpeckleBanner.value) return false
-  if (projectsPanelResult?.value?.activeUser?.projectInvites.length) return false
-
-  return true
-})
-
-const onDismissNewSpeckleBanner = () => {
-  hasDismissedNewSpeckleBanner.value = true
 }
 </script>
