+++ packages/frontend-2/lib/dashboard/graphql/queries.ts
@@ -9,6 +9,7 @@ export const dashboardProjectsPageQuery = graphql(`
           ...DashboardProjectCard_Project
         }
       }
+      ...ProjectsDashboardHeader_User
     }
   }
 `)
