+++ packages/frontend-2/pages/dashboard.vue
@@ -1,5 +1,7 @@
 <template>
   <div class="flex flex-col gap-y-12">
+    <ProjectsDashboardHeader :user="projectsResult?.activeUser || undefined" />
+
     <section>
       <h2 class="text-heading-sm text-foreground-2">Quick start</h2>
       <div class="grid grid-cols-1 md:grid-cols-3 gap-3 pt-5">
